Move serde implementation of `FlatPieces` into submodule
